refactor: extract answer collection actions dropdown to separate component
